Moved shell secondary functions into separate file; Introduced funcs/ directory for organized function storage;
